Cleanup, add documentation, move IslandBuilder
Authored-by: Julian Zucker <julian.zucker@gmail.com>
