+++ src/main/scala/com/evvo/island/EvvoIslandBuilder.scala
@@ -1,0 +1,52 @@
+package com.evvo.island
+
+import akka.actor.{ActorSystem, Props}
+import com.evvo.agent.{CreatorFunction, DeletorFunction, MutatorFunction}
+import com.evvo.island.population.Objective
+
+/**
+  * @param creators   the functions to be used for creating new solutions.
+  * @param mutators   the functions to be used for creating new solutions from current solutions.
+  * @param deletors   the functions to be used for deciding which solutions to delete.
+  * @param objectives the objective functions to maximize.
+  */
+case class EvvoIslandBuilder[Sol]
+(
+  creators: Set[CreatorFunction[Sol]] = Set[CreatorFunction[Sol]](),
+  mutators: Set[MutatorFunction[Sol]] = Set[MutatorFunction[Sol]](),
+  deletors: Set[DeletorFunction[Sol]] = Set[DeletorFunction[Sol]](),
+  objectives: Set[Objective[Sol]] = Set[Objective[Sol]]()
+) {
+  def addCreator(creatorFunc: CreatorFunction[Sol]): EvvoIslandBuilder[Sol] = {
+    this.copy(creators = creators + creatorFunc)
+  }
+
+  def addMutator(mutatorFunc: MutatorFunction[Sol]): EvvoIslandBuilder[Sol] = {
+    this.copy(mutators = mutators + mutatorFunc)
+  }
+
+  def addDeletor(deletorFunc: DeletorFunction[Sol]): EvvoIslandBuilder[Sol] = {
+    this.copy(deletors = deletors + deletorFunc)
+  }
+
+  def addObjective(objective: Objective[Sol])
+  : EvvoIslandBuilder[Sol] = {
+    this.copy(objectives = objectives + objective)
+  }
+
+  def toProps()(implicit system: ActorSystem): Props = {
+    Props(new RemoteEvvoIsland[Sol](
+      creators.toVector,
+      mutators.toVector,
+      deletors.toVector,
+      objectives.toVector))
+  }
+
+  def buildLocalEvvo(): EvolutionaryProcess[Sol] = {
+    new LocalEvvoIsland[Sol](
+      creators.toVector,
+      mutators.toVector,
+      deletors.toVector,
+      objectives.toVector)
+  }
+}
